+++ Magic/CHANGELOG.md
@@ -10,6 +10,7 @@
  - Allow using wands with `craftable: true` in furnace recipes
    NOTE: Smelthing recipes with wand ingredients still doesn't seem to work but I will
          have to look into this further at a later date.
+ - Add `worlds` requirement type
 
 # 10.8.5